SERVER/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
google-generativeai
reportlab
pydantic

# Database + auth
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator